+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxAuxData.hs
@@ -30,6 +30,7 @@ module Cardano.Ledger.Alonzo.TxAuxData (
   hashAlonzoTxAuxData,
   validateAlonzoTxAuxData,
   getAlonzoTxAuxDataScripts,
+  translateAlonzoTxAuxData,
 
   -- * Deprecated
   AuxiliaryData,
@@ -37,7 +38,8 @@ module Cardano.Ledger.Alonzo.TxAuxData (
 where
 
 import Cardano.Crypto.Hash.Class (HashAlgorithm)
-import Cardano.Ledger.Allegra.Scripts
+import Cardano.Ledger.Allegra.Scripts (Timelock, translateTimelock)
+import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..))
 import Cardano.Ledger.Alonzo.Era
 import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..), BinaryPlutus (..), validScript)
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
@@ -235,7 +237,17 @@ type AuxiliaryData era = AlonzoTxAuxData era
 
 instance Crypto c => EraTxAuxData (AlonzoEra c) where
   type TxAuxData (AlonzoEra c) = AlonzoTxAuxData (AlonzoEra c)
+
+  upgradeTxAuxData (AllegraTxAuxData md scripts) =
+    mkMemoized $
+      AlonzoTxAuxDataRaw
+        { atadrMetadata = md
+        , atadrTimelock = translateTimelock <$> scripts
+        , atadrPlutus = mempty
+        }
+
   hashTxAuxData = hashAlonzoTxAuxData
+
   validateTxAuxData = validateAlonzoTxAuxData
 
 hashAlonzoTxAuxData ::
@@ -279,7 +291,7 @@ pattern AlonzoTxAuxData ::
   Map Word64 Metadatum ->
   StrictSeq (Timelock era) ->
   Map Language (NE.NonEmpty BinaryPlutus) ->
-  AuxiliaryData era
+  AlonzoTxAuxData era
 pattern AlonzoTxAuxData {atadMetadata, atadTimelock, atadPlutus} <-
   (getMemoRawType -> AlonzoTxAuxDataRaw atadMetadata atadTimelock atadPlutus)
   where
@@ -287,3 +299,14 @@ pattern AlonzoTxAuxData {atadMetadata, atadTimelock, atadPlutus} <-
       mkMemoized $ AlonzoTxAuxDataRaw {atadrMetadata, atadrTimelock, atadrPlutus}
 
 {-# COMPLETE AlonzoTxAuxData #-}
+
+translateAlonzoTxAuxData ::
+  (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) =>
+  AlonzoTxAuxData era1 ->
+  AlonzoTxAuxData era2
+translateAlonzoTxAuxData AlonzoTxAuxData {atadMetadata, atadTimelock, atadPlutus} =
+  AlonzoTxAuxData
+    { atadMetadata = atadMetadata
+    , atadTimelock = translateTimelock <$> atadTimelock
+    , atadPlutus = atadPlutus
+    }