+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxAuxData.hs
@@ -1,0 +1,22 @@
+{-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+module Cardano.Ledger.Conway.TxAuxData () where
+
+import Cardano.Ledger.Alonzo.TxAuxData (
+  AlonzoTxAuxData,
+  hashAlonzoTxAuxData,
+  validateAlonzoTxAuxData,
+  translateAlonzoTxAuxData
+ )
+import Cardano.Ledger.Conway.Era
+import Cardano.Ledger.Core
+import Cardano.Ledger.Crypto
+
+instance Crypto c => EraTxAuxData (ConwayEra c) where
+  type TxAuxData (ConwayEra c) = AlonzoTxAuxData (ConwayEra c)
+
+  upgradeTxAuxData = translateAlonzoTxAuxData
+
+  hashTxAuxData = hashAlonzoTxAuxData
+
+  validateTxAuxData = validateAlonzoTxAuxData