+++ eras/mary/impl/src/Cardano/Ledger/Mary/TxAuxData.hs
@@ -1,18 +1,10 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE PatternSynonyms #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Mary.TxAuxData (
@@ -22,7 +14,7 @@ where
 
 import Cardano.Ledger.Allegra.TxAuxData
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
-import Cardano.Ledger.Core (EraTxAuxData (..))
+import Cardano.Ledger.Core (EraTxAuxData (..), upgradeScript)
 import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Mary.Era (MaryEra)
 import Cardano.Ledger.Mary.Scripts ()
@@ -34,5 +26,9 @@ import Control.DeepSeq (deepseq)
 
 instance Crypto c => EraTxAuxData (MaryEra c) where
   type TxAuxData (MaryEra c) = AllegraTxAuxData (MaryEra c)
+
+  upgradeTxAuxData (AllegraTxAuxData md scripts) = AllegraTxAuxData md $ upgradeScript <$> scripts
+
   validateTxAuxData _ (AllegraTxAuxData md as) = as `deepseq` all validMetadatum md
+
   hashTxAuxData aux = AuxiliaryDataHash (hashAnnotated aux)