+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxAuxData.hs
@@ -112,7 +112,12 @@ type Metadata era = ShelleyTxAuxData era
 instance Crypto c => EraTxAuxData (ShelleyEra c) where
   type TxAuxData (ShelleyEra c) = ShelleyTxAuxData (ShelleyEra c)
 
+  -- Calling this partial function will result in compilation error, since ByronEra has
+  -- no instance for EraTxOut type class.
+  upgradeTxAuxData = error "It is not possible to translate Byron TxOut with 'upgradeTxOut'"
+
   validateTxAuxData _ (ShelleyTxAuxData m) = all validMetadatum m
+
   hashTxAuxData metadata =
     AuxiliaryDataHash (makeHashWithExplicitProxys (Proxy @c) index metadata)
     where