+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -382,7 +382,16 @@ class
   EraTxAuxData era
   where
   type TxAuxData era = (r :: Type) | r -> era
+
+  -- | Every era, except Shelley, must be able to upgrade a `TxAuxData` from a previous
+  -- era.
+  --
+  -- /Warning/ - Important to note that any memoized binary representation will not be
+  -- preserved, you need to retain underlying bytes you can use `translateEraThroughCBOR`
+  upgradeTxAuxData :: EraTxAuxData (PreviousEra era) => TxAuxData (PreviousEra era) -> TxAuxData era
+
   hashTxAuxData :: TxAuxData era -> AuxiliaryDataHash (EraCrypto era)
+
   validateTxAuxData :: ProtVer -> TxAuxData era -> Bool
 
 type AuxiliaryData era = TxAuxData era
@@ -460,6 +469,9 @@ class
   type Script era = (r :: Type) | r -> era
 
   -- | Every era, except Shelley, must be able to upgrade a `Script` from a previous era.
+  --
+  -- /Warning/ - Important to note that any memoized binary representation will not be
+  -- preserved, you need to retain underlying bytes you can use `translateEraThroughCBOR`
   upgradeScript :: EraScript (PreviousEra era) => Script (PreviousEra era) -> Script era
 
   scriptPrefixTag :: Script era -> BS.ByteString