+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4.0.0
 
+* Add `upgradeTxAucData` function to `EraTxAuxData`
 * Add `upgradeTxOut` function to `EraTxOut`
 * Add `upgradeScript` function to `EraScript`
 * Export: